Make simple_rpc work
Had to improve the Sembast-pretending-to-be-like-mongo abstraction including moving sembast to int keys so they could fit into ObjectId's. Otherwise Sembast uses 20char base 64 keys, where as ObjectIds are 24 char base 16 keys.

EventSource handling works now, even if it's very primitive/hacky, although when the client disconnects it doesn't seem to notify the stream immediately which it should (something is broken).

Removed fromDbJson and toDbJSon from ClassInfo as the dbJson format is per-db specific, so instead just having the model objects make network json which the db's know how to convert to/from.

Fixed the /newMessages getter to actually work, previously it was bound to POST (and broken in other ways).