Refactor handlers.dart to be nicer to generate
Tried to make handlers.dart as small as possible to make it easy to generate.  My experience is generated code is much harder to edit so trying to push as much as possible into library functions instead.

Also added a source_gen dependency since it has a nice(r) way to check for annotations than using the analyzer apis directly.