Make `shorebird generate` work for simple_rpc handlers
It works!  It even revealed a bug (I wasn't previously allowing arguments from the request to be passed to a Stream based handler).

handlers.dart and new/handlers.dart still differ in that the former (manual) uses a separate class for the args.  I think I'm going to go back to that as part of implementing client.dart generation next though.

Once I'm satisfied with the quality of the generated code I will replace the files in gen/ with these new ones which are currently being generated to gen/new/.