Actually use the generated code
Previously we were using the hand-written "gen" code, now the "gen" code is all from `shorebird generate`. This found a few remaining bugs in the generate scripts which I fixed.

This also starts a `shorebird run`, but currently it leaks processes, so it's not useful yet.  Will fix that in the next commit.