Start to sketch out a test executable for the updater lib
The idea here is to build a libupdater.so which can be linked and called from either the flutter engine or flutter embeddings (not sure yet), but that handles all the updater logic for shorebird in a portable fashion.

After investigating the current state of Zig, I decided to use Rust, mostly because the Zig http client is < 1mo old and not available in the version of Zig which was easily installable on my machine.

Rust will likely produce a larger binary than Zig (or C or C++) would have, but should be fine for a demo.

I will need to split the command line "demo" binary out from the main library once this is further along so that the main library doesn't need to depend on the clap crate, etc.